Update testWhiteboard.py
For this new commit, I got rid of any excess features that we did not need in the whiteboard, like changing the pixel size. I updated the ability for the highlighter to work with a yellow marker, but changed the opacity to make it lighter. There is an eraser as well that works on command. The textbox feature is being implemented, and at this step, it still is not 100% working. It will place textboxes, but for EVERY time the left mouse is clicked. That will be fixed soon.